Make the other Authentication controllers invokable
